requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.9.5
aiosignal==1.3.2
anyio==4.8.0
appdirs==1.4.4
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
multidict==6.1.0
pillow==10.3.0
propcache==0.3.0
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
